requirements.txt
fastapi
transformers
uvicorn
torch
numpy
beautifulsoup4
requests
python-dotenv
llama-index
llama-cloud-services
openai
requests
langchain 
langchain-core 
langchain-openai
ipywidgets
safetensors
huggingface-hub==0.36.0